# docs/swagger.yaml
openapi: 3.0.0
info:
  title: BrainyWrite API
  description: API para o website de consultoria BrainyWrite. Gerencia contatos, cotações, agendamentos e newsletter.
  version: 1.0.0
  contact:
    email: contato.brainywrite@gmail.com
servers:
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento
  - url: https://api.brainywrite.co.mz
    description: Servidor de Produção

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    
    Success:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string

    ContactInput:
      type: object
      required: [nome, apelido, email, telefone, assunto, mensagem]
      properties:
        nome:
          type: string
        apelido:
          type: string
        email:
          type: string
          format: email
        telefone:
          type: string
        assunto:
          type: string
        mensagem:
          type: string

    QuoteInput:
      type: object
      required: [name, email, phone, workType, pages, deadline]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        institution:
          type: string
        workType:
          type: string
        category:
          type: string
        pages:
          type: integer
        deadline:
          type: string
          format: date
        academicLevel:
          type: string
        formatting:
          type: string
        language:
          type: string
        title:
          type: string
        description:
          type: string
        hasReferences:
          type: string
          enum: [sim, nao]
        additionalServices:
          type: array
          items:
            type: string
        calculatedPrice:
          type: number

    AppointmentInput:
      type: object
      required: [name, email, phone, service, date, time]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        service:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        message:
          type: string

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

security:
  - bearerAuth: []

paths:
  /api/health:
    get:
      summary: Verificar status da API
      security: []
      responses:
        200:
          description: API está online
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  message:
                    type: string

  /api/contact:
    post:
      summary: Enviar mensagem de contato
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactInput'
      responses:
        200:
          description: Mensagem enviada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/quote:
    post:
      summary: Solicitar cotação
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteInput'
      responses:
        200:
          description: Cotação solicitada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: Erro de validação

  /api/appointment:
    post:
      summary: Agendar consulta
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentInput'
      responses:
        200:
          description: Agendamento realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: Erro de validação

  /api/newsletter:
    post:
      summary: Inscrever na newsletter
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        200:
          description: Inscrição realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
    get:
      summary: Listar inscritos (Admin)
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de inscritos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    email:
                      type: string
                    subscribed_at:
                      type: string
                      format: date-time

  /api/admin/login:
    post:
      summary: Login administrativo
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        200:
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    type: object
        401:
          description: Credenciais inválidas

  /api/admin/stats:
    get:
      summary: Obter estatísticas do dashboard
      security:
        - bearerAuth: []
      responses:
        200:
          description: Estatísticas gerais
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_contacts:
                    type: integer
                  unread_contacts:
                    type: integer
                  total_quotes:
                    type: integer
                  pending_quotes:
                    type: integer

  /api/contacts:
    get:
      summary: Listar contatos recebidos
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de contatos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactInput'

  /api/contacts/{id}/read:
    patch:
      summary: Marcar contato como lido
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Contato atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

  /api/quotes:
    get:
      summary: Listar cotações recebidas
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de cotações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteInput'

  /api/quotes/{id}/status:
    patch:
      summary: Atualizar status da cotação
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, contacted, converted, rejected]
      responses:
        200:
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

  /api/appointments:
    get:
      summary: Listar agendamentos
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de agendamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentInput'

  /api/appointments/{id}/status:
    patch:
      summary: Atualizar status do agendamento
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, cancelled, completed]
      responses:
        200:
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'